+++ src/ReactiveUI.Primitives/Signal/ReadOnlyState{T}.cs
@@ -2,8 +2,6 @@
 // ReactiveUI Association Incorporated licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-#pragma warning disable SA1116, SA1117, SA1204, SA1402, SA1501, SA1611, SA1615, SA1618
-
 namespace ReactiveUI.Primitives.Signals;
 
 /// <summary>
@@ -49,9 +47,11 @@ public sealed class ReadOnlyState<T> : IObservable<T>, IDisposable
     public IObservable<T> Changed => _inner;
 
     /// <summary>
-    /// Executes the Subscribe operation.
+    /// Notifies the provider that an observer is to receive notifications.
     /// </summary>
-    /// <returns>The result.</returns>
+    /// <param name="observer">The object that is to receive notifications.</param>
+    /// <returns>A reference to an interface that allows observers to stop receiving notifications before the provider has
+    /// finished sending them.</returns>
     public IDisposable Subscribe(IObserver<T> observer) => _inner.Subscribe(observer);
 
     /// <summary>
@@ -61,44 +61,5 @@ public sealed class ReadOnlyState<T> : IObservable<T>, IDisposable
     {
         _subscription.Dispose();
         _inner.Dispose();
-    }
-}
-
-/// <summary>
-/// State projection helpers.
-/// </summary>
-public static class StateSignalMixins
-{
-    /// <summary>
-    /// Projects an observable sequence into a read-only state signal.
-    /// </summary>
-    /// <typeparam name="TSource">The source value type.</typeparam>
-    /// <typeparam name="TResult">The projected value type.</typeparam>
-    /// <param name="source">The source sequence.</param>
-    /// <param name="initialValue">The initial projected value.</param>
-    /// <param name="selector">The projection function.</param>
-    /// <returns>A read-only projected state.</returns>
-    public static ReadOnlyState<TResult> ToReadOnlyState<TSource, TResult>(
-        this IObservable<TSource> source,
-        TResult initialValue,
-        Func<TSource, TResult> selector)
-    {
-        if (source == null)
-        {
-            throw new ArgumentNullException(nameof(source));
-        }
-
-        if (selector == null)
-        {
-            throw new ArgumentNullException(nameof(selector));
-        }
-
-        return new ReadOnlyState<TResult>(
-            ReactiveUI.Primitives.Signals.Signal.CreateSafe<TResult>(
-                observer => source.Subscribe(
-                    value => observer.OnNext(selector(value)),
-                    observer.OnError,
-                    observer.OnCompleted)),
-            initialValue);
     }
 }
