+++ src/ReactiveUI.Primitives/Signal/StateSignal{T}.cs
@@ -4,8 +4,6 @@
 
 using ReactiveUI.Primitives;
 
-#pragma warning disable SA1501
-
 namespace ReactiveUI.Primitives.Signals;
 
 /// <summary>
