+++ src/ReactiveUI.Primitives/SignalOperatorParityMixins.cs
@@ -8,8 +8,6 @@ using ReactiveUI.Primitives.Disposables;
 using ReactiveUI.Primitives.Signals;
 using ReactiveUI.Primitives.Signals.Core;
 
-#pragma warning disable SA1107, SA1116, SA1117, SA1501, SA1611, SA1615, SA1618
-
 namespace ReactiveUI.Primitives;
 
 /// <summary>
@@ -20,11 +18,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Prepends a value before the source sequence. Alias of <see cref="Prepend{T}(IObservable{T}, T)"/> using Primitives vocabulary.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to emit before the source.</param>
+    /// <returns>A sequence that emits <paramref name="value"/> before the source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Lead<T>(this IObservable<T> source, T value) => source.Prepend(value);
 
     /// <summary>
     /// Prepends a value before the source sequence.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to emit before the source.</param>
+    /// <returns>A sequence that emits <paramref name="value"/> before the source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Prepend<T>(this IObservable<T> source, T value)
     {
         if (source == null)
@@ -38,6 +46,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Appends a value after the source sequence completes.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to emit after the source completes.</param>
+    /// <returns>A sequence that emits the source values followed by <paramref name="value"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Append<T>(this IObservable<T> source, T value)
     {
         if (source == null)
@@ -51,11 +64,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Prepends a value before the source sequence using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to emit before the source.</param>
+    /// <returns>A sequence that emits <paramref name="value"/> before the source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> StartWith<T>(this IObservable<T> source, T value) => source.Prepend(value);
 
     /// <summary>
     /// Prepends values before the source sequence using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="values">The values to emit before the source.</param>
+    /// <returns>A sequence that emits <paramref name="values"/> before the source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="values"/> is <see langword="null"/>.</exception>
     public static IObservable<T> StartWith<T>(this IObservable<T> source, params T[] values)
     {
         if (source == null)
@@ -74,6 +97,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Prepends values before the source sequence using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="values">The values to emit before the source.</param>
+    /// <returns>A sequence that emits <paramref name="values"/> before the source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="values"/> is <see langword="null"/>.</exception>
     public static IObservable<T> StartWith<T>(this IObservable<T> source, IEnumerable<T> values)
     {
         if (source == null)
@@ -92,22 +120,40 @@ public static partial class LinqMixins
     /// <summary>
     /// Returns the source as an observable. This is an identity adapter for BCL observable sources.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>The supplied source sequence.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> AsObservable<T>(this IObservable<T> source) => source ?? throw new ArgumentNullException(nameof(source));
 
     /// <summary>
     /// Converts an enumerable sequence to a Primitives signal using the System.Reactive conversion name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="values">The values to enumerate.</param>
+    /// <returns>A signal that emits the enumerable values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="values"/> is <see langword="null"/>.</exception>
     public static IObservable<T> ToObservable<T>(this IEnumerable<T> values) => Signal.FromEnumerable(values);
 
     /// <summary>
     /// Converts an enumerable sequence to a Primitives signal using the System.Reactive conversion name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="values">The values to enumerate.</param>
+    /// <param name="cancellationToken">The token used to stop enumeration.</param>
+    /// <returns>A signal that emits the enumerable values until enumeration completes or cancellation is requested.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="values"/> is <see langword="null"/>.</exception>
     public static IObservable<T> ToObservable<T>(this IEnumerable<T> values, CancellationToken cancellationToken) =>
         Signal.FromEnumerable(values, cancellationToken);
 
     /// <summary>
     /// Schedules observer notifications on the supplied scheduler using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="scheduler">The sequencer used to deliver observer notifications.</param>
+    /// <returns>The source sequence when <paramref name="scheduler"/> is immediate; otherwise a sequence observed on the sequencer.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="scheduler"/> is <see langword="null"/>.</exception>
     public static IObservable<T> ObserveOn<T>(this IObservable<T> source, ISequencer scheduler)
     {
         if (source == null)
@@ -131,23 +177,48 @@ public static partial class LinqMixins
     /// <summary>
     /// Alias for <see cref="DelayStart{T}(IObservable{T}, TimeSpan, ISequencer?)"/> using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The delay before subscribing to the source.</param>
+    /// <returns>A sequence that subscribes to the source after <paramref name="dueTime"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DelaySubscription<T>(this IObservable<T> source, TimeSpan dueTime) =>
         source.DelayStart(dueTime, null);
 
     /// <summary>
     /// Alias for <see cref="DelayStart{T}(IObservable{T}, TimeSpan, ISequencer?)"/> using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The delay before subscribing to the source.</param>
+    /// <param name="scheduler">The sequencer used to schedule the delayed subscription.</param>
+    /// <returns>A sequence that subscribes to the source after <paramref name="dueTime"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DelaySubscription<T>(this IObservable<T> source, TimeSpan dueTime, ISequencer? scheduler) =>
         source.DelayStart(dueTime, scheduler);
 
     /// <summary>
     /// Runs a side effect for each source value using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="onNext">The action to invoke for each value.</param>
+    /// <returns>The source values after <paramref name="onNext"/> runs.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="onNext"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Do<T>(this IObservable<T> source, Action<T> onNext) => source.Tap(onNext);
 
     /// <summary>
-    /// Runs side effects for source notifications using the System.Reactive operator name.
+    /// Invokes actions for each element in the observable sequence, for error notifications, and for successful
+    /// completion.
     /// </summary>
+    /// <typeparam name="T">The type of the elements in the source sequence.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="onNext">Action to invoke for each element in the observable sequence.</param>
+    /// <param name="onError">Action to invoke upon exceptional termination of the observable sequence.</param>
+    /// <param name="onCompleted">Action to invoke upon graceful termination of the observable sequence.</param>
+    /// <returns>The source sequence with the side-effecting behavior applied.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/>, <paramref name="onNext"/>, <paramref name="onError"/>, or <paramref
+    /// name="onCompleted"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Do<T>(this IObservable<T> source, Action<T> onNext, Action<Exception> onError, Action onCompleted)
     {
         if (source == null)
@@ -191,12 +262,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Alias for <see cref="Rescue{T}(IObservable{T}, Func{Exception, IObservable{T}})"/> using the System.Reactive operator name.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="handler">The function that creates the recovery sequence for an error.</param>
+    /// <returns>A sequence that continues with the handler result after an error.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="handler"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Catch<T>(this IObservable<T> source, Func<Exception, IObservable<T>> handler) =>
         source.Rescue(handler);
 
     /// <summary>
     /// Ignores all source values and only forwards terminal messages.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that forwards only error and completion notifications.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> IgnoreValues<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -210,12 +290,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits the supplied value if the source completes without values.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that emits <see langword="default"/> when the source is empty.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DefaultIfEmpty<T>(this IObservable<T> source) =>
         source.DefaultIfEmpty(default!);
 
     /// <summary>
     /// Emits the supplied value if the source completes without values.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="defaultValue">The value to emit when the source is empty.</param>
+    /// <returns>A sequence that emits <paramref name="defaultValue"/> when the source is empty.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DefaultIfEmpty<T>(this IObservable<T> source, T defaultValue)
     {
         if (source == null)
@@ -229,12 +318,25 @@ public static partial class LinqMixins
     /// <summary>
     /// Suppresses duplicate keys according to the comparer.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <typeparam name="TKey">The key type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="keySelector">The function that selects the comparison key.</param>
+    /// <returns>A sequence containing the first value for each observed key.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="keySelector"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DistinctBy<T, TKey>(this IObservable<T> source, Func<T, TKey> keySelector) =>
         source.DistinctBy(keySelector, null);
 
     /// <summary>
     /// Suppresses duplicate keys according to the comparer.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <typeparam name="TKey">The key type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="keySelector">The function that selects the comparison key.</param>
+    /// <param name="comparer">The comparer used to identify duplicate keys.</param>
+    /// <returns>A sequence containing the first value for each observed key.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="keySelector"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DistinctBy<T, TKey>(this IObservable<T> source, Func<T, TKey> keySelector, IEqualityComparer<TKey>? comparer)
     {
         if (source == null)
@@ -253,12 +355,25 @@ public static partial class LinqMixins
     /// <summary>
     /// Suppresses adjacent duplicate keys according to the comparer.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <typeparam name="TKey">The key type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="keySelector">The function that selects the comparison key.</param>
+    /// <returns>A sequence with adjacent duplicate keys removed.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="keySelector"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DistinctUntilChangedBy<T, TKey>(this IObservable<T> source, Func<T, TKey> keySelector) =>
         source.DistinctUntilChangedBy(keySelector, null);
 
     /// <summary>
     /// Suppresses adjacent duplicate keys according to the comparer.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <typeparam name="TKey">The key type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="keySelector">The function that selects the comparison key.</param>
+    /// <param name="comparer">The comparer used to compare adjacent keys.</param>
+    /// <returns>A sequence with adjacent duplicate keys removed.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="keySelector"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DistinctUntilChangedBy<T, TKey>(this IObservable<T> source, Func<T, TKey> keySelector, IEqualityComparer<TKey>? comparer)
     {
         if (source == null)
@@ -297,6 +412,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits values while the predicate remains true, then completes.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that determines whether to keep taking values.</param>
+    /// <returns>A sequence that emits the leading values that satisfy <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<T> TakeWhile<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (source == null)
@@ -338,6 +458,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Skips values while the predicate remains true, then mirrors the remaining source.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that determines whether to keep skipping values.</param>
+    /// <returns>A sequence that emits values after the leading values that satisfy <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<T> SkipWhile<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (source == null)
@@ -372,11 +497,23 @@ public static partial class LinqMixins
     /// <summary>
     /// Projects each source value to an inner signal and concatenates all inner values.
     /// </summary>
+    /// <typeparam name="TSource">The source value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="selector">The function that projects each source value to an inner sequence.</param>
+    /// <returns>A sequence containing the concatenated inner values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="selector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> Bind<TSource, TResult>(this IObservable<TSource> source, Func<TSource, IObservable<TResult>> selector) => source.SelectMany(selector);
 
     /// <summary>
     /// Projects each source value to an inner signal and concatenates all inner values.
     /// </summary>
+    /// <typeparam name="TSource">The source value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="selector">The function that projects each source value to an inner sequence.</param>
+    /// <returns>A sequence containing the concatenated inner values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="selector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> SelectMany<TSource, TResult>(this IObservable<TSource> source, Func<TSource, IObservable<TResult>> selector)
     {
         if (source == null)
@@ -395,6 +532,14 @@ public static partial class LinqMixins
     /// <summary>
     /// Projects each source value to an inner signal and maps outer/inner values with a result selector.
     /// </summary>
+    /// <typeparam name="TSource">The source value type.</typeparam>
+    /// <typeparam name="TCollection">The inner value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="collectionSelector">The function that projects each source value to an inner sequence.</param>
+    /// <param name="resultSelector">The function that combines source and inner values.</param>
+    /// <returns>A sequence containing selected outer/inner combinations.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="collectionSelector"/> or <paramref name="resultSelector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> SelectMany<TSource, TCollection, TResult>(
         this IObservable<TSource> source,
         Func<TSource, IObservable<TCollection>> collectionSelector,
@@ -416,6 +561,10 @@ public static partial class LinqMixins
     /// <summary>
     /// Counts the source values as an <see cref="int"/>.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that emits the number of source values when the source completes.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<int> Count<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -429,6 +578,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Counts source values that satisfy the predicate as an <see cref="int"/>.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that identifies values to count.</param>
+    /// <returns>A sequence that emits the matching value count when the source completes.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<int> Count<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (predicate == null)
@@ -447,6 +601,10 @@ public static partial class LinqMixins
     /// <summary>
     /// Counts the source values as an <see cref="long"/>.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that emits the number of source values when the source completes.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<long> LongCount<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -460,6 +618,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Counts source values that satisfy the predicate as an <see cref="long"/>.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that identifies values to count.</param>
+    /// <returns>A sequence that emits the matching value count when the source completes.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<long> LongCount<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (predicate == null)
@@ -478,6 +641,10 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits true when any value is present.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that emits whether the source produced any values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> Any<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -491,6 +658,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits true when any value satisfies the predicate.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that tests each value.</param>
+    /// <returns>A sequence that emits whether any source value satisfies <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> Any<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (source == null)
@@ -509,6 +681,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits true when every value satisfies the predicate.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that tests each value.</param>
+    /// <returns>A sequence that emits whether every source value satisfies <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> All<T>(this IObservable<T> source, Func<T, bool> predicate)
     {
         if (source == null)
@@ -553,12 +730,23 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits true when the source contains the requested value.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to locate.</param>
+    /// <returns>A sequence that emits whether the source contains <paramref name="value"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> Contains<T>(this IObservable<T> source, T value) =>
         source.Contains(value, null);
 
     /// <summary>
     /// Emits true when the source contains the requested value.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="value">The value to locate.</param>
+    /// <param name="comparer">The comparer used to compare source values.</param>
+    /// <returns>A sequence that emits whether the source contains <paramref name="value"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> Contains<T>(this IObservable<T> source, T value, IEqualityComparer<T>? comparer)
     {
         comparer ??= EqualityComparer<T>.Default;
@@ -568,17 +756,32 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits true when the source completes without values.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence that emits whether the source completed without values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<bool> IsEmpty<T>(this IObservable<T> source) => source.Any().Map(hasValue => !hasValue);
 
     /// <summary>
     /// Emits values from source after delaying subscription by the due time.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The delay before subscribing to the source.</param>
+    /// <returns>A sequence that subscribes to the source after <paramref name="dueTime"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DelayStart<T>(this IObservable<T> source, TimeSpan dueTime) =>
         source.DelayStart(dueTime, null);
 
     /// <summary>
     /// Emits values from source after delaying subscription by the due time.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The delay before subscribing to the source.</param>
+    /// <param name="scheduler">The sequencer used to schedule the delayed subscription.</param>
+    /// <returns>A sequence that subscribes to the source after <paramref name="dueTime"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> DelayStart<T>(this IObservable<T> source, TimeSpan dueTime, ISequencer? scheduler)
     {
         if (source == null)
@@ -598,12 +801,23 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits only the most recent value after the quiet period elapses.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The quiet period before emitting the latest value.</param>
+    /// <returns>A sequence that emits the latest value after each quiet period.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Throttle<T>(this IObservable<T> source, TimeSpan dueTime) =>
         source.Throttle(dueTime, null);
 
     /// <summary>
     /// Emits only the most recent value after the quiet period elapses.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="dueTime">The quiet period before emitting the latest value.</param>
+    /// <param name="scheduler">The sequencer used to schedule quiet-period timers.</param>
+    /// <returns>A sequence that emits the latest value after each quiet period.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<T> Throttle<T>(this IObservable<T> source, TimeSpan dueTime, ISequencer? scheduler)
     {
         if (source == null)
@@ -650,12 +864,25 @@ public static partial class LinqMixins
     /// <summary>
     /// Emits the latest source value whenever the sampling period ticks.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="period">The interval between sampling ticks.</param>
+    /// <returns>A sequence that emits the latest source value on each sampling tick.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="period"/> is less than <see cref="TimeSpan.Zero"/>.</exception>
     public static IObservable<T> Sample<T>(this IObservable<T> source, TimeSpan period) =>
         source.Sample(period, null);
 
     /// <summary>
     /// Emits the latest source value whenever the sampling period ticks.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="period">The interval between sampling ticks.</param>
+    /// <param name="scheduler">The sequencer used to schedule sampling ticks.</param>
+    /// <returns>A sequence that emits the latest source value on each sampling tick.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="period"/> is less than <see cref="TimeSpan.Zero"/>.</exception>
     public static IObservable<T> Sample<T>(this IObservable<T> source, TimeSpan period, ISequencer? scheduler)
     {
         if (source == null)
@@ -677,12 +904,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Annotates values with their scheduler timestamp.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence containing each value with its timestamp.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<Moment<T>> Timestamp<T>(this IObservable<T> source) =>
         source.Timestamp(null);
 
     /// <summary>
     /// Annotates values with their scheduler timestamp.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="scheduler">The sequencer that supplies timestamps.</param>
+    /// <returns>A sequence containing each value with its timestamp.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<Moment<T>> Timestamp<T>(this IObservable<T> source, ISequencer? scheduler)
     {
         if (source == null)
@@ -697,12 +933,21 @@ public static partial class LinqMixins
     /// <summary>
     /// Annotates each value with the elapsed scheduler time since the previous value.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A sequence containing each value with its elapsed interval since the previous value.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<TimeInterval<T>> TimeInterval<T>(this IObservable<T> source) =>
         source.TimeInterval(null);
 
     /// <summary>
     /// Annotates each value with the elapsed scheduler time since the previous value.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="scheduler">The sequencer that supplies timestamps.</param>
+    /// <returns>A sequence containing each value with its elapsed interval since the previous value.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static IObservable<TimeInterval<T>> TimeInterval<T>(this IObservable<T> source, ISequencer? scheduler)
     {
         if (source == null)
@@ -732,18 +977,42 @@ public static partial class LinqMixins
     /// <summary>
     /// Combines latest values from both sources. Alias for latest-fusion vocabulary.
     /// </summary>
+    /// <typeparam name="TLeft">The left value type.</typeparam>
+    /// <typeparam name="TRight">The right value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="left">The left sequence.</param>
+    /// <param name="right">The right sequence.</param>
+    /// <param name="selector">The function that combines the latest values.</param>
+    /// <returns>A sequence containing selected latest-value combinations.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="left"/>, <paramref name="right"/>, or <paramref name="selector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> ZipLatest<TLeft, TRight, TResult>(this IObservable<TLeft> left, IObservable<TRight> right, Func<TLeft, TRight, TResult> selector) =>
         left.CombineLatest(right, selector);
 
     /// <summary>
     /// Alias for <see cref="ZipLatest{TLeft, TRight, TResult}(IObservable{TLeft}, IObservable{TRight}, Func{TLeft, TRight, TResult})"/>.
     /// </summary>
+    /// <typeparam name="TLeft">The left value type.</typeparam>
+    /// <typeparam name="TRight">The right value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="left">The left sequence.</param>
+    /// <param name="right">The right sequence.</param>
+    /// <param name="selector">The function that combines the latest values.</param>
+    /// <returns>A sequence containing selected latest-value combinations.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="left"/>, <paramref name="right"/>, or <paramref name="selector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> FuseLatest<TLeft, TRight, TResult>(this IObservable<TLeft> left, IObservable<TRight> right, Func<TLeft, TRight, TResult> selector) =>
         left.ZipLatest(right, selector);
 
     /// <summary>
     /// Waits for both sources to complete and emits one value from their last elements when both produced at least one value.
     /// </summary>
+    /// <typeparam name="TLeft">The left value type.</typeparam>
+    /// <typeparam name="TRight">The right value type.</typeparam>
+    /// <typeparam name="TResult">The result value type.</typeparam>
+    /// <param name="left">The left sequence.</param>
+    /// <param name="right">The right sequence.</param>
+    /// <param name="selector">The function that combines the final values.</param>
+    /// <returns>A sequence that emits one selected value after both sources complete.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="left"/>, <paramref name="right"/>, or <paramref name="selector"/> is <see langword="null"/>.</exception>
     public static IObservable<TResult> ForkJoin<TLeft, TRight, TResult>(this IObservable<TLeft> left, IObservable<TRight> right, Func<TLeft, TRight, TResult> selector)
     {
         if (left == null)
@@ -779,6 +1048,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Awaits the first source value.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with the first source value.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
+    /// <exception cref="InvalidOperationException">The source completes without producing a value.</exception>
     public static Task<T> FirstAsync<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -797,6 +1071,10 @@ public static partial class LinqMixins
     /// <summary>
     /// Awaits the first source value, returning a default value when the source is empty.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with the first source value, or <see langword="default"/> when the source is empty.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<T> FirstOrDefaultAsync<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -815,6 +1093,11 @@ public static partial class LinqMixins
     /// <summary>
     /// Awaits the first source value, returning a default value when the source is empty.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="defaultValue">The value to return when the source is empty.</param>
+    /// <returns>A task that completes with the first source value, or <paramref name="defaultValue"/> when the source is empty.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<T> FirstOrDefaultAsync<T>(this IObservable<T> source, T defaultValue)
     {
         if (source == null)
@@ -833,11 +1116,22 @@ public static partial class LinqMixins
     /// <summary>
     /// Awaits source completion and returns the last value produced by the source.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with the final source value.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
+    /// <exception cref="InvalidOperationException">The source completes without producing a value.</exception>
     public static Task<T> ToTask<T>(this IObservable<T> source) => source.ToTask(CancellationToken.None);
 
     /// <summary>
     /// Awaits source completion and returns the last value produced by the source.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="cancellationToken">The token used to cancel the task and dispose the subscription.</param>
+    /// <returns>A task that completes with the final source value.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
+    /// <exception cref="InvalidOperationException">The source completes without producing a value.</exception>
     [System.Diagnostics.CodeAnalysis.SuppressMessage(
         "Major Code Smell",
         "S1541:Methods and properties should not be too complex",
@@ -910,59 +1204,107 @@ public static partial class LinqMixins
     /// <summary>
     /// Identity helper that keeps source-compatible <c>FirstAsync().ToTask()</c> migrations compiling.
     /// </summary>
+    /// <typeparam name="T">The task result type.</typeparam>
+    /// <param name="task">The task to return.</param>
+    /// <returns>The supplied task.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="task"/> is <see langword="null"/>.</exception>
     public static Task<T> ToTask<T>(this Task<T> task) => task ?? throw new ArgumentNullException(nameof(task));
 
     /// <summary>
     /// Awaits the source count as a task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with the number of source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<int> CountAsync<T>(this IObservable<T> source) =>
         source.Count().ToTask();
 
     /// <summary>
     /// Awaits the source count as a task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="cancellationToken">The token used to cancel the task.</param>
+    /// <returns>A task that completes with the number of source values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<int> CountAsync<T>(this IObservable<T> source, CancellationToken cancellationToken) =>
         source.Count().ToTask(cancellationToken);
 
     /// <summary>
     /// Awaits the source predicate count as a task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that identifies values to count.</param>
+    /// <returns>A task that completes with the matching value count.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static Task<int> CountAsync<T>(this IObservable<T> source, Func<T, bool> predicate) =>
         source.Count(predicate).ToTask();
 
     /// <summary>
     /// Awaits the source predicate count as a task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that identifies values to count.</param>
+    /// <param name="cancellationToken">The token used to cancel the task.</param>
+    /// <returns>A task that completes with the matching value count.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static Task<int> CountAsync<T>(this IObservable<T> source, Func<T, bool> predicate, CancellationToken cancellationToken) =>
         source.Count(predicate).ToTask(cancellationToken);
 
     /// <summary>
     /// Awaits whether any value is present.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with whether the source produced any values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<bool> AnyAsync<T>(this IObservable<T> source) =>
         source.Any().ToTask();
 
     /// <summary>
     /// Awaits whether any value is present.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="cancellationToken">The token used to cancel the task.</param>
+    /// <returns>A task that completes with whether the source produced any values.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<bool> AnyAsync<T>(this IObservable<T> source, CancellationToken cancellationToken) =>
         source.Any().ToTask(cancellationToken);
 
     /// <summary>
     /// Awaits whether any value matches a predicate.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that tests each value.</param>
+    /// <returns>A task that completes with whether any source value satisfies <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static Task<bool> AnyAsync<T>(this IObservable<T> source, Func<T, bool> predicate) =>
         source.Any(predicate).ToTask();
 
     /// <summary>
     /// Awaits whether any value matches a predicate.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <param name="predicate">The function that tests each value.</param>
+    /// <param name="cancellationToken">The token used to cancel the task.</param>
+    /// <returns>A task that completes with whether any source value satisfies <paramref name="predicate"/>.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> or <paramref name="predicate"/> is <see langword="null"/>.</exception>
     public static Task<bool> AnyAsync<T>(this IObservable<T> source, Func<T, bool> predicate, CancellationToken cancellationToken) =>
         source.Any(predicate).ToTask(cancellationToken);
 
     /// <summary>
     /// Collects all values into an array task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with all source values in an array.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<T[]> CollectArrayAsync<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -984,6 +1326,10 @@ public static partial class LinqMixins
     /// <summary>
     /// Collects all values into a list task.
     /// </summary>
+    /// <typeparam name="T">The value type.</typeparam>
+    /// <param name="source">The source sequence.</param>
+    /// <returns>A task that completes with all source values in a list.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="source"/> is <see langword="null"/>.</exception>
     public static Task<IList<T>> CollectListAsync<T>(this IObservable<T> source)
     {
         if (source == null)
@@ -1051,8 +1397,13 @@ public static partial class LinqMixins
     }
 
     /// <summary>
-    /// Creates a generic value from an integer range item.
+    /// Converts an integer value to the specified numeric type.
     /// </summary>
+    /// <remarks>Uses boxing and unboxing to perform the conversion. The generic type parameter is expected to
+    /// be validated by the caller.</remarks>
+    /// <typeparam name="T">The target numeric type.</typeparam>
+    /// <param name="value">The integer value to convert.</param>
+    /// <returns>The value converted to type <typeparamref name="T"/>.</returns>
     [System.Diagnostics.CodeAnalysis.SuppressMessage(
         "Major Code Smell",
         "S4018:Generic methods should provide type parameters",
@@ -1060,8 +1411,11 @@ public static partial class LinqMixins
     private static T CreateRangeValue<T>(int value) => (T)(object)value;
 
     /// <summary>
-    /// Creates a range-backed array for task terminal fast paths.
+    /// Creates an array of sequential values from the specified range signal.
     /// </summary>
+    /// <typeparam name="T">The element type of the array.</typeparam>
+    /// <param name="range">The range signal specifying the start value and count.</param>
+    /// <returns>An array containing sequential values from the range start to start + count - 1.</returns>
     [System.Diagnostics.CodeAnalysis.SuppressMessage(
         "Major Code Smell",
         "S4018:Generic methods should provide type parameters",
@@ -1089,8 +1443,13 @@ public static partial class LinqMixins
     }
 
     /// <summary>
-    /// Creates a range-backed list for task terminal fast paths.
+    /// Creates a list of values from the specified range signal.
     /// </summary>
+    /// <remarks>Optimized for integer types by directly incrementing values. For other types, uses
+    /// <c>CreateRangeValue</c> to generate each element.</remarks>
+    /// <typeparam name="T">The type of elements to create in the list.</typeparam>
+    /// <param name="range">The range signal containing the start value and count.</param>
+    /// <returns>A list containing the generated range values.</returns>
     [System.Diagnostics.CodeAnalysis.SuppressMessage(
         "Major Code Smell",
         "S4018:Generic methods should provide type parameters",
